requirements.txt
djangorestframework==3.3.1
Django==1.8.5

